requirements.txt
Django~=4.2.8
djangorestframework~=3.14.0
openai~=1.3.9
requests~=2.31.0